Fixed GitHub issue #1.
This involved rewriting the .usm-env script to get closer to pure POSIX
sh compatability (as measured against the sh.exe in msysgit).

I also rewrote part of the USM script so that the .usm-env script is
embedded in the file without having to do all sorts of nasty escape
quotations.